Refactor content_directory into files
