+++ crates/upnp-serve/src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod connection_manager;
+pub mod content_directory;
